Refactor StopList
