+++ frontend/components/lines/Single/index.tsx
@@ -15,7 +15,6 @@ import SelectActivePatternGroup from '@/components/lines/SelectActivePatternGrou
 import StopList from '@/components/lines/StopList';
 import { useLinesSingleContext } from '@/contexts/LinesSingle.context';
 import { useProfileContext } from '@/contexts/Profile.context';
-import { useState } from 'react';
 
 import styles from './styles.module.css';
 
@@ -29,9 +28,6 @@ export default function Component() {
 
 	const profileContext = useProfileContext();
 	const linesSingleContext = useLinesSingleContext();
-
-	const [selectedStop, setSelectedStop] = useState(null);
-	const [selectedStopSequence, setSelectedStopSequence] = useState(null);
 
 	//
 	// B. Handle actions
@@ -70,7 +66,7 @@ export default function Component() {
 			{linesSingleContext.data.active_pattern_group ? (
 				<Section withGap={false} withTopPadding={false} withChildrenPadding>
 					<LineMap />
-					<StopList selectedStop={selectedStop} setSelectedStop={setSelectedStop} setSelectedStopSequence={setSelectedStopSequence} />
+					<StopList />
 				</Section>
 			) : (
 				<Section withGap={false} withTopPadding={false} withChildrenPadding>
