+++ frontend/components/lines/SingleStopFullContent/styles.module.css
@@ -1,0 +1,53 @@
+.label {
+    color: var(--color-system-border-200);
+    font-size: 10px;
+    font-weight: 600;
+    font-style: italic;
+    padding-bottom: 8px;
+}
+.scheduledList, .realtimeList {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    color:var(--color-system-text-100);
+    font-weight: 700;
+}
+.scheduledList>div{
+    display: flex;
+    gap: 15px;
+    align-items: center;
+}
+.realtimeList {
+    gap: 2px;
+    color:var(--color-realtime-100)
+}
+.timesList{
+    display: flex;
+    gap: 10px;
+    align-items: center;
+    padding-bottom: 23px;
+}
+.buttons button{
+    display:flex;
+    background-color: var(--color-system-background-200);
+    color: var(--color-system-text-200);
+    align-items: center;
+    font-size: 14px;
+    padding: 8px 10px;
+    gap: 2px;
+    border-radius: 7px;
+}
+
+.buttons {
+    display: flex;
+    gap: 15px;
+}
+.facilityIcons {
+    padding-top:10px;
+    padding-bottom:20px;
+    display:flex;
+    gap:8px;
+}
+.facilityIcons>svg {
+    height: 24px;
+}
